Fix test for groups route
Debug: The issues was highlighted when moving some default configs to
conftest and making the drop and create table only per session. Since
the group test to read groups of users was using hardcoded values for
user starting from 1,2,.. this worked only when each test module the
entire table was recreated which meant the auto generated sequence ids
for groups started from 1,2... however when making creation of table
only session scoped means if any group were added in other tests the
sequence would be past 1,2,.. when it reaches group testing.

Fix: Use the dynamic values of group id that was returned when creating
groups